// types/supabase.ts
export interface definitions {
  revue_subscribers: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    email: string;
    first_name?: string;
    last_name?: string;
    created_at?: string;
    /** the time at which the subscriber was synced to convertkit */
    convertkit_synced_at?: string;
  };
}